ADD: Modularización
Se realiza la separación de las entidades.

issue #20
