--- src/schemas/mockups/Comentarios.json
@@ -1,7 +1,0 @@
-[
-    {
-        "id": "1",
-        "nombre": "Jorge Mendez Ortegs",
-        "cuerpo": "Comentario loco"
-    }
-]
