--- src/schemas/mockups/Profesores.json
@@ -1,5 +1,0 @@
-{
-    "id": "1",
-    "nombre": "Jorge Mendez",
-    "nacionalidad": "Mexicana"
-}
