--- src/schemas/mockups/index.js
@@ -1,9 +1,0 @@
-import MockCursos from "./Cursos.json";
-import MockComentarios from "./Comentarios.json";
-import MockProfesores from "./Profesores.json";
-
-export {
-    MockComentarios,
-    MockCursos,
-    MockProfesores,
-};
