--- src/schemas/mockupsCasual/Comentarios.js
@@ -1,16 +1,0 @@
-import Casual from "casual";
-
-let contador = 1;
-const DATA = [];
-
-do {
-    contador += 1;
-    const SCHEMA = {
-        id: Casual.uuid,
-        cuerpo: Casual.description,
-        nombre: Casual.full_name,
-    };
-    DATA.push(SCHEMA);
-} while (contador < 11);
-
-export default DATA;
