--- src/schemas/mockupsCasual/index.js
@@ -1,9 +1,0 @@
-import CasualCursos from "./Cursos";
-import CasualComentarios from "./Comentarios";
-import CasualProfesores from "./Profesores";
-
-export {
-    CasualComentarios,
-    CasualCursos,
-    CasualProfesores,
-};
