+++ src/schemas/resolvers/index.js
@@ -1,0 +1,3 @@
+import resolver from "./resolver";
+
+export default resolver;
