+++ src/schemas/schemas/index.js
@@ -1,0 +1,7 @@
+import Curso from "./Curso";
+import Profesor from "./Profesor";
+
+export {
+    Curso,
+    Profesor,
+};
